fix(firmware): close 32-bit overflow in deploy_parse_header + 32-bit CI coverage
deploy_parse_header summed DEPLOY_HDR_BYTES + chunk_len in size_t; on the
32-bit target a crafted chunk_len near UINT32_MAX wraps the sum below msg_len
and passes the "payload fits" gate — the same overflow class fixed for the
manifest path in 4ef438a, one function up in the same file. Promote the add to
uint64_t and add adversarial header-parse checks.

The 64-bit host build can't observe size_t-width bugs, so the regression test
passed either way there. Add a `make test-m32` target (builds the suite 32-bit,
skips cleanly where the toolchain can't) and wire it into the firmware CI job
with gcc-multilib — verified red->green against the reverted fix on a 32-bit
build.

+++ firmware/src/deploy.c
@@ -27,7 +27,10 @@ bool deploy_parse_header(const uint8_t *msg, size_t msg_len, deploy_hdr_t *out) 
     if (out->part < DEPLOY_PART_MANIFEST || out->part > DEPLOY_PART_MODEL) {
         return false;
     }
-    if ((size_t)DEPLOY_HDR_BYTES + out->chunk_len > msg_len) {
+    // uint64 add: size_t is 32-bit on the target, so a crafted chunk_len near UINT32_MAX would
+    // wrap a 32-bit sum below msg_len and pass this gate, breaking the "payload fits" guarantee
+    // every downstream consumer relies on. Same guard the reassembly path uses below.
+    if ((uint64_t)DEPLOY_HDR_BYTES + out->chunk_len > msg_len) {
         return false;  // declared payload doesn't fit the message
     }
     return true;
